+++ src/main/java/roomescape/controller/api/MemberController.java
@@ -8,8 +8,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.dto.request.MemberRequest;
-import roomescape.dto.response.MemberResponse;
+import roomescape.controller.dto.request.MemberRequest;
+import roomescape.controller.dto.response.MemberResponse;
 import roomescape.service.MemberService;
 
 @RestController
