+++ src/main/java/roomescape/controller/api/ReservationController.java
@@ -13,9 +13,9 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.config.annotation.AuthMember;
-import roomescape.dto.request.ReservationRequest;
-import roomescape.dto.response.MyReservationResponse;
-import roomescape.dto.response.ReservationResponse;
+import roomescape.controller.dto.request.ReservationRequest;
+import roomescape.controller.dto.response.MyReservationResponse;
+import roomescape.controller.dto.response.ReservationResponse;
 import roomescape.entity.Member;
 import roomescape.service.ReservationService;
