+++ src/main/java/roomescape/controller/api/ThemeController.java
@@ -10,8 +10,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.dto.request.ThemeRequest;
-import roomescape.dto.response.ThemeResponse;
+import roomescape.controller.dto.request.ThemeRequest;
+import roomescape.controller.dto.response.ThemeResponse;
 import roomescape.service.ThemeService;
 
 @RestController
