+++ src/main/java/roomescape/controller/dto/response/LoginResponse.java
@@ -1,4 +1,4 @@
-package roomescape.dto.response;
+package roomescape.controller.dto.response;
 
 import roomescape.entity.Member;
