+++ src/main/java/roomescape/repository/JpaMemberRepository.java
@@ -1,4 +1,4 @@
-package roomescape.repository.jpa;
+package roomescape.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import roomescape.entity.Member;
