+++ src/main/java/roomescape/service/ThemeService.java
@@ -3,10 +3,10 @@ package roomescape.service;
 import java.time.LocalDate;
 import java.util.List;
 import org.springframework.stereotype.Service;
-import roomescape.dto.request.ThemeRequest;
+import roomescape.controller.dto.request.ThemeRequest;
 import roomescape.entity.Theme;
 import roomescape.exception.custom.DuplicatedException;
-import roomescape.repository.jpa.JpaThemeRepository;
+import roomescape.repository.JpaThemeRepository;
 
 @Service
 public class ThemeService {
