+++ src/test/java/roomescape/dto/request/LoginRequestTest.java
@@ -4,7 +4,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import roomescape.dto.request.LoginRequest;
+import roomescape.controller.dto.request.LoginRequest;
 import roomescape.exception.custom.InvalidInputException;
 
 class LoginRequestTest {
