+++ src/test/java/roomescape/dto/request/ReservationRequestTest.java
@@ -5,7 +5,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import java.time.LocalDate;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import roomescape.dto.request.ReservationRequest;
+import roomescape.controller.dto.request.ReservationRequest;
 import roomescape.exception.custom.InvalidInputException;
 
 class ReservationRequestTest {
