+++ src/test/java/roomescape/dto/request/TimeRequestTest.java
@@ -4,7 +4,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import roomescape.dto.request.TimeRequest;
+import roomescape.controller.dto.request.TimeRequest;
 import roomescape.exception.custom.InvalidInputException;
 
 class TimeRequestTest {
